requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
albumentations>=1.3.0
Pillow>=10.0.0

# Data Science & Numerics
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Experiment Tracking & Configuration
wandb>=0.15.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Training Utilities (optional but recommended)
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0

# Progress Bars & Logging
tqdm>=4.65.0
loguru>=0.7.0

# Web App & Demo
streamlit>=1.25.0
gradio>=3.40.0

# Video Processing
moviepy>=1.0.3
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Data Handling
pycocotools>=2.0.6
lxml>=4.9.0
xmltodict>=0.13.0

# Jupyter & Development
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0
notebook>=7.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0

# GPU Monitoring (optional)
gpustat>=1.1.0
py3nvml>=0.2.7
